requirements.txt
paho-mqtt>=1.6.1
aprslib>=0.7.2
setproctitle>=1.3.0
